Move module array to tools index file and refactor tool palette
